fix: 강의실 삭제 아이콘 위치 이동
