+++ src/component/ui/class/AdminTopBar.jsx
@@ -1,8 +1,11 @@
-import { NavLink, useParams } from "react-router-dom";
+import { useState, useContext } from "react";
+import { NavLink, useParams, useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import PropTypes from "prop-types";
 import Delete from "../../img/icon/bin.svg";
 import Share from "../../img/icon/share.svg";
+import api from "../../api/api";
+import { UsersContext } from "../../contexts/usersContext";
 
 const Container = styled.div`
   width: 100%;
@@ -82,8 +85,78 @@ const Icon = styled.img`
   }
 `;
 
+const DeleteModalOverlay = styled.div`
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.5);
+    z-index: 1000;
+`;
+
+const DeleteModal = styled.div`
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    background-color: #fff;
+    min-width: 350px;
+    width: 40%;
+    height: 30%;
+    border-radius: 12px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+`;
+
+const CloseIcon = styled.span`
+    position: absolute;
+    top: 20px;
+    right: 20px;
+    font-size: 28px;
+    cursor: pointer;
+`;
+
+const ModalTitle = styled.h2`
+    margin: 0;
+    font-size: 20px;
+    font-weight: normal;
+`;
+
+const YesButton = styled.button`
+    margin-top: 40px;
+    width: 100px;
+    height: 40px;
+    border: none;
+    border-radius: 24px;
+    background-color: var(--main-color);
+    color: #fff;
+    font-size: 16px;
+    cursor: pointer;
+    &:hover {
+        opacity: 0.9;
+    }
+`;
+
 const AdminTopBar = ({ activeTab }) => {
   const { courseId } = useParams();
+  const { user } = useContext(UsersContext);
+  const navigate = useNavigate();
+  // 모달 관련 상태
+  const [showDeleteModal, setShowDeleteModal] = useState(false);
+
+  const handleDeleteLecture = async (courseId) => {
+    console.log('Deleting course with ID:', courseId);
+    try {
+        await api.delete(`/courses/${courseId}?userId=${user.userId}`);
+        navigate("/class/list");
+    } catch (error) {
+        console.error('강의실 삭제 중 오류 발생:', error);
+    }
+};
 
   return (
     <Container>
@@ -112,9 +185,12 @@ const AdminTopBar = ({ activeTab }) => {
 
         <IconContainer>
           <Icon 
-            className="delete-icon material-icons" 
+            className="material-icons" 
             src={Delete} 
             alt="delete icon" 
+            onClick={() => {
+              setShowDeleteModal(true);
+            }}
           />
           <Icon 
             className="material-icons" 
@@ -123,6 +199,15 @@ const AdminTopBar = ({ activeTab }) => {
           />
         </IconContainer>
       </NavbarContent>
+      {showDeleteModal && (
+        <DeleteModalOverlay>
+          <DeleteModal>
+            <CloseIcon onClick={() => setShowDeleteModal(false)}>×</CloseIcon>
+            <ModalTitle>강의실을 삭제하시겠습니까?</ModalTitle>
+            <YesButton onClick={() => handleDeleteLecture(courseId)}>예</YesButton>
+          </DeleteModal>
+        </DeleteModalOverlay>
+      )}
     </Container>
   );
 };
